[FLINK-31621][table] fix review.

+++ flink-table/flink-table-common/src/main/java/org/apache/flink/table/functions/BuiltInFunctionDefinitions.java
@@ -229,7 +229,7 @@ public final class BuiltInFunctionDefinitions {
                                     Arrays.asList("haystack", "needle"),
                                     Arrays.asList(
                                             logical(LogicalTypeRoot.ARRAY), ARRAY_ELEMENT_ARG)))
-                    .outputTypeStrategy(nullableIfArgs(explicit(DataTypes.BIGINT())))
+                    .outputTypeStrategy(nullableIfArgs(explicit(DataTypes.INT())))
                     .runtimeClass(
                             "org.apache.flink.table.runtime.functions.scalar.ArrayPositionFunction")
                     .build();
